requirements.txt
python
pycryptodome